// src/content/projectiles.generated.ts
export type ProjectileId =
  | 'pistol'
  | 'assault_rifle'
  | 'shotgun'
  | 'submachine_gun'
  | 'gauss_gun'
  | 'plasma_rifle'
  | 'plasma_minigun'
  | 'pulse_gun'
  | 'ion_rifle'
  | 'ion_minigun'
  | 'ion_cannon'
  | 'shrinkifier'
  | 'blade_gun'
  | 'spider_plasma'
  | 'plasma_cannon'
  | 'splitter_gun'
  | 'plague_spreader'
  | 'rainbow_gun'
  | 'fire_bullets';
// ...
export const PROJECTILE_BY_TYPE_ID: Record<number, ProjectileId> = {
  [1]: 'pistol',
  [2]: 'assault_rifle',
  [3]: 'shotgun',
  [5]: 'submachine_gun',
  [6]: 'gauss_gun',
  [9]: 'plasma_rifle',
  [11]: 'plasma_minigun',
  [19]: 'pulse_gun',
  [21]: 'ion_rifle',
  [22]: 'ion_minigun',
  [23]: 'ion_cannon',
  [24]: 'shrinkifier',
  [25]: 'blade_gun',
  [26]: 'spider_plasma',
  [28]: 'plasma_cannon',
  [29]: 'splitter_gun',
  [41]: 'plague_spreader',
  [43]: 'rainbow_gun',
  [45]: 'fire_bullets',
};